+++ tests/cli/test_yolo_format.py
@@ -51,7 +51,7 @@ class YoloIntegrationScenarios(TestCase):
                 "-f",
                 "yolo",
                 "--",
-                "--save-images",
+                "--save-media",
             )
 
             parsed_dataset = Dataset.import_from(export_dir, format="yolo")
@@ -73,7 +73,7 @@ class YoloIntegrationScenarios(TestCase):
             run(self, "import", "-p", test_dir, "-f", "mot_seq", mot_dir)
 
             yolo_dir = osp.join(test_dir, "yolo_dir")
-            run(self, "export", "-p", test_dir, "-o", yolo_dir, "-f", "yolo", "--", "--save-images")
+            run(self, "export", "-p", test_dir, "-o", yolo_dir, "-f", "yolo", "--", "--save-media")
 
             parsed_dataset = Dataset.import_from(yolo_dir, format="yolo")
             compare_datasets(self, target_dataset, parsed_dataset)
@@ -91,7 +91,10 @@ class YoloIntegrationScenarios(TestCase):
                         Bbox(4.0, 5.0, 2.0, 2.0, label=15),
                         Bbox(5.5, 6, 2, 2, label=22),
                     ],
-                )
+                ),
+                DatasetItem(
+                    id="2007_000002", subset="test", media=Image(data=np.ones((10, 20, 3)))
+                ),
             ],
             categories=[label.name for label in VOC.make_voc_categories()[AnnotationType.label]],
         )
@@ -118,57 +121,11 @@ class YoloIntegrationScenarios(TestCase):
                 "-o",
                 yolo_dir,
                 "--",
-                "--save-images",
+                "--save-media",
             )
 
             parsed_dataset = Dataset.import_from(yolo_dir, format="yolo")
             compare_datasets(self, target_dataset, parsed_dataset, require_media=True)
-
-    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
-    def test_can_ignore_non_supported_subsets(self):
-        source_dataset = Dataset.from_iterable(
-            [
-                DatasetItem(
-                    id="img1",
-                    subset="test",
-                    media=Image(data=np.ones((10, 20, 3))),
-                    annotations=[Bbox(1.0, 2.0, 1.0, 1.0, label=0)],
-                ),
-                DatasetItem(
-                    id="img2",
-                    subset="train",
-                    media=Image(data=np.ones((10, 5, 3))),
-                    annotations=[Bbox(3.0, 1.0, 2.0, 1.0, label=1)],
-                ),
-            ],
-            categories=[str(i) for i in range(4)],
-        )
-
-        target_dataset = Dataset.from_iterable(
-            [
-                DatasetItem(
-                    id="img2",
-                    subset="train",
-                    media=Image(data=np.ones((10, 5, 3))),
-                    annotations=[Bbox(3.0, 1.0, 2.0, 1.0, label=1)],
-                )
-            ],
-            categories=[str(i) for i in range(4)],
-        )
-
-        with TestDir() as test_dir:
-            dataset_dir = osp.join(test_dir, "dataset_dir")
-            source_dataset.save(dataset_dir, save_media=True)
-
-            proj_dir = osp.join(test_dir, "proj")
-            run(self, "create", "-o", proj_dir)
-            run(self, "import", "-p", proj_dir, "-f", "datumaro", dataset_dir)
-
-            yolo_dir = osp.join(test_dir, "yolo_dir")
-            run(self, "export", "-p", proj_dir, "-o", yolo_dir, "-f", "yolo", "--", "--save-images")
-
-            parsed_dataset = Dataset.import_from(yolo_dir, format="yolo")
-            compare_datasets(self, target_dataset, parsed_dataset)
 
     @mark_requirement(Requirements.DATUM_GENERAL_REQ)
     def test_can_delete_labels_from_yolo_dataset(self):